pyproject.toml
[project]
name = "plpq"
version = "0.1.0"
description = ""
authors = [
    {name = "Louis Maddox", email = "louismmx@gmail.com"},
]
dependencies = [
    "polars-lts-cpu>=0.19.19",
    "huggingface-hub>=0.19.4",
    "fsspec>=2023.10.0",
    "pandas>=2.1.3",
    "pyarrow>=14.0.1",
    "hf-transfer>=0.1.4",
    "httpx>=0.25.2",
    "aria2p[tui]>=0.11.3",
]
requires-python = ">=3.10,<3.13"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
test = [
    "pytest>=7.4.3",
    "pytest-random-order>=1.1.0",
    "pytest-xdist>=3.5.0",
]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
